refactor(components/ui/table): keep default patterns
