+++ src/presentation/components/ui/table/index.ts
@@ -4,8 +4,8 @@ import { Cell } from './cell'
 import { Footer } from './footer'
 import { Head } from './head'
 import { Header } from './header'
+import { Root } from './root'
 import { Row } from './row'
-import { Table as Root } from './table'
 
 export const Table = {
 	Root,
